Create route for content controller.
We need a specific route for the static video content controller, a
requirement which this commit fulfills. All static content serves will be
under the 'content' route prefix, this specifically encompasses static
content that is meant to be hidden behind authentication barriers.

This commit is a part of task LANGLEAP-78 for story LANGLEAP-37.