Add internal X-Accel paths to configuration.
Nginx has a unique XSendfile implementation. Where as Apache and Lighttpd
allow you to specify a path and then define your whitelist/blacklist
separately, Nginx requires that you define a fixed location configuration
that is only accessible to Nginx itself. This "internal location" then is
alias'd to a specific directory where all files are served from.

Therefore, for Nginx's unique case, we must maintain a list of the
locations where videos should be served from.

This commit is a part of task LANGLEAP-78, under story LAGNLEAP-37.