Add controller for video clips.
THe serving of video clips to the client must be handled via a controlled
method that allows for authentication and authorization capabilities. The
addition of the controller in this commit allows for the fulfillment of
this requirement by allowing static video content to be served in a
controlled manner.

The commit contains the controller, the associated tests, and an
abstraction class. The abstraction class basically acts as a factory for
the SplFileInfo class. This allows us to swap in a mock class for the
purposes of testing.

This commit completes task LANGLEAP-78 and is associated with story
LANGLEAP-37.