Fetching questions from the API.
Going to use the AngularJS framework for this. Currently it fetches the
initial question on-boot (i.e. initial page load) from a sample API.

Still need to work on requesting further questions as well as sending the
results back to the API. Given the time, will also add some unit tests to
the ng code.

[ci skip]

LANGLEAP-71