Support CSRF-filtering for API requests.
This implementation broadcasts the CSRF token in the response header for
each page load, thus allowing a client to easily retrieve and use it for
any request that requires it. Furthermore, API clients may send the CSRF
token via the request header rather than in the request body. The
combination of these two takes this authentication procedure out-of-band
of the request content, thus helping to make our API cleaner.

Needed for LANGLEAP-71