CSRF header inject fails on BinaryFileResponse
The issue was brought up by Michael wherein attempting to inject the CSRF
token in the header of a BinaryFileResponse (i.e. a call to the
VideoContentController) results in a 500 error due to the
BinaryFileResponse class not being a child of `Illuminate\Http\Response`,
hence the missing `header()` method.

The changes include verifying that the `$response` object is a child of
`Illuminate\Http\Response` as well as adding this case to the test
coverage.

LANGLEAP-71