Complete quiz flow
The overall UX "flow" of the quiz is complete. Quiz questions are answered
by selecting the appropriate radio button.

As specified by stakeholder desires, questions are submitted as soon as a
selection is made. Once answered, the user is shown whether they were
correct or incorrect, immediately. They can then proceed to the following
question. After answering the final question, they are then shown a modal
dialog listing the result of their quiz. From the modal dialog, they can
then continue onto the following video.

How well this flow functions remains to be tested; currently needs to be
integrated with the API.

LANGLEAP-71