Move to plain REST controller; fix comparisons
The `ApiQuizController` is no longer a resource controller, as that didn't
make much sense for how it was being used. Therefore, it is now a basic
RESTful controller accepting POST and PUT calls on the index.

There was an issue with incorrect comparisons between selected input and
the answer, because JSON integers are regarded as strings. Therefore, we
have to homogenize the variable types and perform a strict comparison.

LANGLEAP-72