Refactor test and impl. for `update` method.
Refactored and increased the test coverage for the `update` method. In
addition to testing for valid changes, we now check that changes to
another user is prevented and that sending in an empty request won't
modify the user. Tests also now verify all attributes of the User model
after requests.

Also updated the implementation to account for the changes with the User
model in commit 0f3273b6d6669101c166c2ee762a283fa0ae527a.

LANGLEAP-84