Refactor test and improve test functionality.
Refactored the test for the `destroy` method and added test case for
attempts to delete another user.

Updated some of the existing test cases to improve their functionality by
using an actual second user instance. Also for the
`testFailsWhenTryingToUpdateAnotherUser` method, added assert statements
to ensure that no changes are actually made in the database.

LANGLEAP-84