Clean up and refactor code.
Cleaned up the code for the `User` model; added annotation for authorship.

Refactored AuthController, simplifying if-logic paths. Modified behaviour
of the `getLogout` method because the action shouldn't result in a JSON
response in accordance with the overarching behaviour of the other
methods and with its intended purpose. It's suppose to redirect the user
to a particular page. Unless we're making an API for a front-end client
here, in which case the other methods are incorrect.

LANGLEAP-90