Fix issue with level ID not refreshing on save.
In the table specifications, the default level ID of a media is defined in
the table schema. An issue arises when we create a new Episode instance,
save it without specifying a level ID, and then attempt to use the same
instance without refreshing it. The Eloquent ORM attempts to save some
database queries by reusing the existing attributes, which lacks the level
ID that was added during the save transaction by the database. Therefore,
anything that attempts to use the `level_id` attribute would encounter a
null value.

The fix was to simply refresh the instance reference from the database.

Also reverted the additional input value in the associated test case.

LANGLEAP-125